refactor: organization -> center 로 변경
