+++ app/src/main/java/com/konkuk/hackathon/feature/signup/center/CenterSignUpViewModel.kt
@@ -1,6 +1,5 @@
-package com.konkuk.hackathon.feature.signup.organization
+package com.konkuk.hackathon.feature.signup.center
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -9,10 +8,10 @@ import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 
 @HiltViewModel
-class OrganizationSignUpViewModel @Inject constructor() : ViewModel() {
+class CenterSignUpViewModel @Inject constructor() : ViewModel() {
 
-    private val _uiState = MutableStateFlow(OrganizationSignUpUiState())
-    val uiState: StateFlow<OrganizationSignUpUiState>
+    private val _uiState = MutableStateFlow(CenterSignUpUiState())
+    val uiState: StateFlow<CenterSignUpUiState>
         get() = _uiState.asStateFlow()
 
     fun updateAllTermsAccepted(checked: Boolean) {
